requirements.txt
yfinance
apache-airflow=3.1.5
kafka-python
psycopg2-binary
pandas
seaborn